+++ .dockerignore
@@ -1,0 +1,2 @@
+transmission
+__pycache__
