+++ config.ini
@@ -1,0 +1,24 @@
+[ByrBTBot]
+byrbt-url = https://byr.pt/
+username = krf573
+passwd = 1154844702
+transmission-user = user
+transmission-passwd = 123456
+max-torrent = 20
+;all size in G
+max-torrent-total-size = 1024
+torrent-max-size = 512
+torrent-min-size = 1
+
+[BitTorrent]
+;bit-torrent-host = 10.112.186.186
+;bit-torrent-port = 9091
+;bit-torrent-username = admin
+;bit-torrent-password = admin
+;bit-torrent-download-path = /downloads
+bit-torrent-download-path = /.bt/
+bit-torrent-host = 10.112.239.245
+bit-torrent-port = 9091
+bit-torrent-username = smy
+bit-torrent-password = smy091819
+
