+++ docker-compose.yml
@@ -1,0 +1,28 @@
+version: "3"
+services:
+  transmission:
+    image: linuxserver/transmission:3.00-r5-ls123
+    container_name: transmission
+    environment:
+      - PUID=1000
+      - PGID=1000
+      - TZ=Asia/Shanghai
+      - TRANSMISSION_WEB_HOME=/combustion-release/
+      - USER=admin
+      - PASS=admin
+    volumes:
+      - ./transmission/data:/config
+      - ./transmission/downloads:/downloads
+      - ./transmission/watch:/watch
+    ports:
+      - "9091:9091"
+      - "51413:51413"
+      - "51413:51413/udp"
+    restart: unless-stopped
+    network_mode: host
+  bot:
+    build:
+      context: .
+    image: smyyan/byrbt-bot-transmission
+    restart: unless-stopped
+    network_mode: host
