+++ login.py
@@ -1,0 +1,74 @@
+# -*- encoding: utf-8 -*-
+"""
+@File    : utils.py
+@Time    : 2021/11/20 16:56
+@Author  : smy
+@Email   : smyyan@foxmail.com
+@Software: PyCharm
+"""
+
+import requests
+from bs4 import BeautifulSoup
+from PIL import Image
+from io import BytesIO
+import pickle
+import time
+import os
+from utils.decaptcha import DeCaptcha
+
+decaptcha = DeCaptcha()
+decaptcha.load_model('utils/captcha_classifier.pkl')
+
+
+class LoginTool:
+
+    def __init__(self, config):
+        self.config = config
+        self.try_count = 5
+        self.base_url = str(config.get_bot("byrbt-url"))
+        self.login_url = self.get_url('login.php')
+        self.headers = {
+            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/60.0.3112.113 Safari/537.36'}
+
+    def get_url(self, url):
+        return self.base_url + url
+
+    def load_cookie(self):
+        byrbt_cookies = None
+        if os.path.exists('ByrbtCookies.pickle'):
+            print('find ByrbtCookies.pickle, loading cookies')
+            read_path = open('ByrbtCookies.pickle', 'rb')
+            byrbt_cookies = pickle.load(read_path)
+        else:
+            print('not find ByrbtCookies.pickle, get cookies...')
+            byrbt_cookies = self.login()
+
+        return byrbt_cookies
+
+    def login(self):
+        session = requests.session()
+        for i in range(5):
+            login_content = session.get(self.login_url)
+            login_soup = BeautifulSoup(login_content.text, 'lxml')
+
+            img_url = self.base_url + login_soup.select('#nav_block > form > table > tr:nth-of-type(3) img')[0].attrs['src']
+            img_file = Image.open(BytesIO(session.get(img_url).content))
+
+            captcha_text = decaptcha.decode(img_file)
+
+            login_res = session.post(self.get_url('takelogin.php'),
+                                     headers=self.headers,
+                                     data=dict(username=str(self.config.get_bot("username")),
+                                               password=str(self.config.get_bot("passwd")),
+                                               imagestring=captcha_text,
+                                               imagehash=img_url.split('=')[-1]))
+            if '最近消息' in login_res.text:
+                cookies = {}
+                for k, v in session.cookies.items():
+                    cookies[k] = v
+
+                with open('ByrbtCookies.pickle', 'wb') as f:
+                    pickle.dump(cookies, f)
+                return cookies
+
+            time.sleep(1)
