+++ requirements.txt
@@ -1,10 +1,12 @@
-# Python 3.7.4 推荐用anaconda装
-beautifulsoup4==4.8.0
-lxml==4.4.1
-numpy==1.17.2
-pillow==6.2.0
-# pillow==7.0.0 好像6.2.2以下版本有漏洞？
-requests==2.22.0
+# python==3.7.12
+# pip==21.3.1
+
+beautifulsoup4==4.10.0
+numpy==1.21.4
+pillow==8.4.0
 scikit-learn==0.21.3
-# scikit-learn==0.22.1 可以使用captcha_classifier_sklearn0.22.1.pkl模型
-scipy==1.3.1
+requests==2.26.0
+lxml==4.6.4
+scipy==1.7.2
+six==1.16.0
+transmission-rpc==3.3.0
