+++ start_bot.sh
@@ -1,0 +1,2 @@
+mkdir -p ./transmission/data ./transmission/downloads ./transmission/watch
+docker-compose up -d
